+++ packages/components/src/DataList/DataList.css.d.ts
@@ -3,6 +3,7 @@ declare const styles: {
   readonly "titleContainer": string;
   readonly "headerFilters": string;
   readonly "headerTitles": string;
+  readonly "batchSelectContainer": string;
   readonly "headerBatchSelect": string;
   readonly "listItem": string;
   readonly "active": string;
@@ -13,6 +14,7 @@ declare const styles: {
   readonly "selected": string;
   readonly "filtering": string;
   readonly "filteringSpinner": string;
+  readonly "srOnly": string;
 };
 export = styles;